restore CertRequest in aea.helpers.base
