+++ tests/test_docs/test_cli_vs_programmatic_aeas/programmatic_aea.py
@@ -27,7 +27,6 @@ from typing import cast
 from aea.aea import AEA
 from aea.aea_builder import AEABuilder
 from aea.configurations.base import ConnectionConfig
-from aea.crypto.certificates import CertRequest
 from aea.crypto.fetchai import FetchAICrypto
 from aea.crypto.helpers import (
     PRIVATE_KEY_PATH_SCHEMA,
@@ -35,6 +34,7 @@ from aea.crypto.helpers import (
     make_certificate,
 )
 from aea.crypto.wallet import Wallet
+from aea.helpers.base import CertRequest
 from aea.identity.base import Identity
 from aea.protocols.base import Protocol
 from aea.registries.resources import Resources
